Chess/requirements.txt
Pillow==12.2.0
pygame==2.6.1
